% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_trim.R
\name{mean_trim}
\alias{mean_trim}
\title{Return the mean of a vector after trimming the smallest and largest values from the vector.}
\usage{
mean_trim(x)
}
\arguments{
\item{x}{Vector of numerics}
}
\description{
Return the mean of a vector after trimming the smallest and largest values from the vector.
}
\examples{
mean_trim(c(1,2,3,4,5,6,7,8,9,10))
}
