% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pythag.R
\name{pythag}
\alias{pythag}
\title{Calculate the length of the third side of a right triangle}
\usage{
pythag(side1, side2, hypotenuse = TRUE)
}
\arguments{
\item{side1}{length of one side of right triangle}

\item{side2}{length of second known side}

\item{hypotenuse}{Default = TRUE if calculated side is hypotenuse}
}
\description{
Calculate the length of the third side of a right triangle
}
\examples{
pythag(side1 = 3, side2 = 4, hypotenuse = TRUE)
pythag(side1 = 3, side2 = 5, hypotenuse = FALSE)
}
